backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic[email]==2.*
pydantic-settings==2.*
sqlalchemy==2.*
alembic==1.*
bcrypt==4.*
PyJWT==2.*
cryptography==44.*
celery[redis]==5.*
redis==5.*
imap-tools==1.*
httpx==0.28.*
beautifulsoup4==4.*
google-auth==2.*
google-auth-oauthlib==1.*
